remove deprecated fields from v1 types
